po: Realized what 'array postamble' is for
It is actually a bitmap explaining how the serialized array elements
make up a potentially sparse array. Most arrays are just dense, and so
have a string of 1 bits the length of the array. But script arrays have
a fixed size, and only the scripts actually used are serialized. So they
have a small serialized array and a bitmap explaining which actual
elements those scripts should occupy.