Get rid of `many` in the decoder
Many allows failing at any point during the decoding, and disguises the
problem that actually occurred. Since we want to parse the entire input,
it is insead replaced by a generic error about extra bytes.

Instead, a combinator that checks for EOF and runs a parser if there are
more bytes will give an error from the parser about what actually
failed, which is better.