c:d: Flesh out dat creation
Figured out that one of the numbers in the footer is the total length of
the string entries in the file table. This is irrelevant to us, but
important to the game (it probably pre-allocates storage space for all
the names).

Added a miscellaneous field to the file information, either entry or
tree version. In original files, this just stores the actual pointer to
the file/directory's name, so is garbage data. Since it doesn't matter,
I decided to store a CRC32 of the original bytes, which could be
validated later on archives built with this tool.