Revise internal implementation of directory tree
Switch to ordered containers, and use a case insensitive string wrapper
for the keys to mimic the filesystem more closely. It _seems_ like ToEE
actually cares somewhat about the order entries occur in, because
rebuilding ToEE1.dat with the hash map tree didn't work (seemed to not
find some fonts), while the ordered map seems to produce a file that
works.